PIED-11/15/17
Content section was added, although only the Product Owner's email was put for now. The About Me text is up, although a page still needs to be created for About Me so that I can link it. The reason why it is important has been stated in about two sentences, though it could be edited depending on how you feel. A Logo was added but I feel we can make a better one that stands out more. I'm keeping the react site link at the bottom as a reference for the About Me page.