PIED-16
Added in a second page by adding in several classes. There is an About Me section using a black navigation box on the top and a short About Us written section now. Contacts are on both pages. https://www.techomoro.com/how-to-create-a-multi-page-website-with-react-in-5-minutes/ This link should help with creating numerous pages in ReactJS.