PIED-51 Finished
Finished with the night/day mode button, made the image myself for the sake of time. It's a beta design but it's functional for now. I'd say this is completed. Also I took the liberty to fix the about us page mistake, Abner is now correctly listed as the Scrum Master.