PIED-84 Filters Top Section
It's not 1:1 like in the wireframe, but it should be at the very least functional, which is the most important thing. Be sure to use npm install before testing since I had to install for the slider. Also Abner, look at the school year section in FilterBar.js and the subject section for help on your part.