PIED-84 Create Post Page
Here's what I have so far for Create Post. I don't know how to fix the checkmark placement on the boxes, and the selection for years only selects one for now since I didn't end up having time yet to look into that. Also the professor and date range items are pre-set for now due to Abner not implementing his filters yet and because I'm awaiting how we're going to be connecting this to the back-end.